only remove existing files
git-svn-id: http://www.kaltura.org/kalorg/html5video/trunk/packager@583 b58a29cf-3064-46da-94c6-1c29cc75c8e5

+++ packager.php
@@ -71,8 +71,10 @@ function pakageClassList($packageName, $html5PlayerClassList, $html5PlayerStyleL
 	$rootFileFolder = '/kaltura-html5-player';
 
 	// Remove the old zip if present
-	if( !unlink( $filename ) ){
-		exit( "error could not remove old package: " . $packageName . '.zip');
+	if( is_file( $filename ) ){
+		if( !unlink( $filename ) ){
+			exit( "error could not remove old package: " . $packageName . '.zip');
+		}
 	}
 
 	if ($zip->open($filename, ZIPARCHIVE::CREATE)!==TRUE) {
